+++ src/components/Order/OrderMain/OrderMain.scss
@@ -8,7 +8,6 @@
   .concertHall {
     display: grid;
     grid-template-rows: 1fr 8fr;
-    background-color: $Gray33;
 
     .stage {
       @include flexCenter;
@@ -25,119 +24,16 @@
         background-color: beige;
       }
     }
-
-    .seatTable {
-      display: grid;
-      grid-template-columns: 1fr 30fr;
-      grid-template-rows: 1fr 20fr;
-      gap: 2px;
-      padding: 2%;
-
-      .seatColumn {
-        display: grid;
-        grid-template-columns: repeat(30, 1fr);
-        color: white;
-
-        .column {
-          @include flexCenter;
-          font-size: 0.7rem;
-        }
-      }
-
-      .seatRow {
-        display: grid;
-        grid-template-rows: repeat(20, 1fr);
-        color: white;
-
-        .row {
-          @include flexCenter;
-          font-size: 0.7rem;
-        }
-      }
-
-      .seatList {
-        display: grid;
-        &.column30 {
-          grid-template-columns: repeat(30, 1fr);
-        }
-        &.row20 {
-          grid-template-rows: repeat(20, 1fr);
-        }
-        gap: 1px;
-
-        .seat {
-          width: 19px;
-          height: 19px;
-          border: 1px solid $Gray33;
-
-          &.selected {
-            background-color: $Red;
-            &:hover {
-              background-color: rgba(229, 8, 21, 0.7);
-              cursor: pointer;
-            }
-          }
-
-          &.reserved {
-            background-color: $Black;
-          }
-
-          &.gradeSR {
-            background-color: rgb(242, 159, 6);
-            &:hover {
-              background-color: rgb(242, 159, 6, 0.7);
-              cursor: pointer;
-            }
-          }
-
-          &.gradeR {
-            background-color: rgba(102, 51, 153);
-            &:hover {
-              background-color: rgba(102, 51, 153, 0.7);
-              cursor: pointer;
-            }
-          }
-
-          &.gradeS {
-            background-color: rgba(70, 130, 180);
-            &:hover {
-              background-color: rgba(70, 130, 180, 0.7);
-              cursor: pointer;
-            }
-          }
-
-          &.gradeA {
-            background-color: rgba(153, 153, 153);
-            &:hover {
-              background-color: rgba(153, 153, 153, 0.7);
-              cursor: pointer;
-            }
-          }
-        }
-      }
-    }
   }
 
   .selectSeatInfo {
     display: grid;
     grid-template-rows: 1fr 3fr 1fr 1fr;
-    background-color: $Gray33;
 
     .text {
       @include flexCenter;
       font-size: 2rem;
       color: white;
-    }
-
-    .selectSeatList {
-      margin: 10px;
-      padding: 20px;
-      border: 2px solid white;
-      font-size: 1.1rem;
-
-      .selectSeat {
-        color: white;
-      }
     }
 
     .totalPrice {
