+++ src/components/Order/OrderMain/SelectedSeatList/SelectedSeatList.scss
@@ -1,0 +1,12 @@
+@import '/src/styles/mixin.scss';
+
+.selectedSeatList {
+  margin: 10px;
+  padding: 20px;
+  border: 2px solid white;
+  font-size: 1.1rem;
+
+  .selectSeat {
+    color: white;
+  }
+}
