+++ src/pages/Order/Order.scss
@@ -11,6 +11,7 @@
     grid-template-rows: 1fr 3fr 1fr;
     width: 1000px;
     height: 800px;
+    background-color: $Gray33;
 
     .orderFooter {
       grid-column: span 2;
